fix: fold basetab code into rhtab files
