--- elements/rh-tabs/BaseTab.css
@@ -1,57 +1,0 @@
-:host {
-  display: flex;
-  flex: none;
-}
-
-:host([vertical]) [part="text"] {
-  max-width: 100%;
-  overflow-wrap: break-word;
-}
-
-[hidden] {
-  display: none !important;
-}
-
-slot[name="icon"] {
-  display: block;
-}
-
-button {
-  margin: 0;
-  font-family: inherit;
-  font-size: 100%;
-  border: 0;
-  position: relative;
-  display: flex;
-  flex: 1;
-  text-decoration: none;
-  cursor: pointer;
-}
-
-button:before,
-button:after {
-  position: absolute;
-  inset: 0;
-  content: "";
-  border-style: solid;
-  padding: 0;
-  margin: 0;
-  background-color: transparent;
-}
-
-button:before {
-  pointer-events: none;
-}
-
-:host([fill]) button {
-  flex-basis: 100%;
-  justify-content: center;
-}
-
-:host(:disabled) button {
-  pointer-events: none;
-}
-
-:host([aria-disabled="true"]) button {
-  cursor: default;
-}
