--- elements/rh-tabs/BaseTabPanel.css
@@ -1,7 +1,0 @@
-:host {
-  display: block;
-}
-
-:host([hidden]) {
-  display: none;
-}
