--- elements/rh-tabs/BaseTabs.css
@@ -1,74 +1,0 @@
-:host {
-  display: block;
-}
-
-[part="tabs-container"] {
-  position: relative;
-  display: flex;
-  overflow: hidden;
-}
-
-[part="tabs-container"]:before {
-  position: absolute;
-  right: 0;
-  bottom: 0;
-  left: 0;
-  border-style: solid;
-}
-
-[part="tabs"],
-[part="panels"] {
-  display: block;
-}
-
-[part="tabs"] {
-  scrollbar-width: none;
-  position: relative;
-  max-width: 100%;
-  overflow-x: auto;
-}
-
-[part="tabs-container"]:before,
-[part="tabs"]:before,
-button:before {
-  position: absolute;
-  right: 0;
-  bottom: 0;
-  left: 0;
-  content: "";
-  border-style: solid;
-}
-
-[part="tabs"]:before,
-button:before {
-  top: 0;
-}
-
-button,
-[part="tabs"]:before {
-  border: 0;
-}
-
-button {
-  flex: none;
-  line-height: 1;
-  opacity: 1;
-}
-
-button:before {
-  border-block-start-width: 0;
-}
-
-button:nth-of-type(1) {
-  margin-inline-end: 0;
-  translate: 0 0;
-}
-
-button:nth-of-type(2) {
-  margin-inline-start: 0;
-  translate: 0 0;
-}
-
-button:disabled {
-  pointer-events: none;
-}
