+++ elements/rh-tabs/rh-tab-panel.css
@@ -7,7 +7,8 @@
   padding-inline: var(--_panels-overflow-padding, var(--rh-space-4xl, 64px));
 }
 
-:host([hidden]) {
+:host([hidden]),
+[hidden] {
   display: none !important;
 }
