+++ elements/rh-tabs/rh-tab-panel.ts
@@ -1,9 +1,8 @@
-import { html } from 'lit';
+import { LitElement, html } from 'lit';
 import { customElement } from 'lit/decorators/custom-element.js';
-
 import { classMap } from 'lit/directives/class-map.js';
 
-import { BaseTabPanel } from './BaseTabPanel.js';
+import { getRandomId } from '@patternfly/pfe-core/functions/random.js';
 
 import { colorContextConsumer, type ColorTheme } from '../../lib/context/color/consumer.js';
 
@@ -16,7 +15,7 @@ import styles from './rh-tab-panel.css';
  *
  */
 @customElement('rh-tab-panel')
-export class RhTabPanel extends BaseTabPanel {
+export class RhTabPanel extends LitElement {
   static readonly version = '{{version}}';
 
   static readonly styles = [styles];
@@ -26,10 +25,32 @@ export class RhTabPanel extends BaseTabPanel {
    */
   @colorContextConsumer() private on?: ColorTheme;
 
+  #internals = this.attachInternals();
+
+  connectedCallback() {
+    super.connectedCallback();
+    this.id ||= getRandomId('pf-tab-panel');
+    this.hidden ??= true;
+    this.#internals.role = 'tabpanel';
+
+    /*
+     To make it easy for screen reader users to navigate from a tab
+     to the beginning of content in the active tabpanel, the tabpanel
+     element has tabindex="0" to include the panel in the page Tab sequence.
+     It is recommended that all tabpanel elements in a tab set are focusable
+     if there are any panels in the set that contain content where the first
+     element in the panel is not focusable.
+     https://www.w3.org/WAI/ARIA/apg/example-index/tabs/tabs-automatic
+    */
+    this.tabIndex = 0;
+  }
+
   render() {
     const { on = '' } = this;
     return html`
-      <div id="rhds-container" class="${classMap({ [on]: !!on })}">${super.render()}</div>
+      <div id="rhds-container" class="${classMap({ [on]: !!on })}">
+        <slot></slot>
+      </div>
     `;
   }
 }
