+++ elements/rh-tabs/rh-tab.css
@@ -1,3 +1,73 @@
+:host {
+  display: flex;
+  flex: none;
+}
+
+:host([vertical]) [part="text"] {
+  max-width: 100%;
+  overflow-wrap: break-word;
+}
+
+[hidden] {
+  display: none !important;
+}
+
+slot[name="icon"] {
+  display: block;
+}
+
+button {
+  margin: 0;
+  font-family: inherit;
+  font-size: 100%;
+  border: 0;
+  position: relative;
+  display: flex;
+  flex: 1;
+  text-decoration: none;
+  cursor: pointer;
+  align-items: center;
+  background-color: inherit;
+  color: var(--_button-text-color);
+  padding-block-start: var(--rh-tabs-link-padding-block-start, var(--rh-space-lg, 16px));
+  padding-inline-start: var(--rh-tabs-link-padding-inline-start, var(--rh-space-2xl, 32px));
+  padding-block-end: var(--rh-tabs-link-padding-block-end, var(--rh-space-lg, 16px));
+  padding-inline-end: var(--rh-tabs-link-padding-inline-end, var(--rh-space-2xl, 32px));
+  max-width: 200px;
+  max-height: 100%;
+}
+
+button:before,
+button:after {
+  position: absolute;
+  inset: 0;
+  content: "";
+  border-style: solid;
+  padding: 0;
+  margin: 0;
+  background-color: transparent;
+}
+
+button:before {
+  pointer-events: none;
+  border: 0 solid transparent;
+}
+
+:host([fill]) button {
+  flex-basis: 100%;
+  justify-content: center;
+}
+
+:host(:disabled) button {
+  pointer-events: none;
+}
+
+:host([aria-disabled="true"]) button {
+  cursor: default;
+}
+
+/* START RHDS */
+
 #rhds-container {
   display: flex;
   width: 100%;
@@ -72,28 +142,10 @@
       );
 }
 
-/* Button Resets */
-button {
-  display: flex;
-  align-items: center;
-  background-color: inherit;
-  color: var(--_button-text-color);
-  padding-block-start: var(--rh-tabs-link-padding-block-start, var(--rh-space-lg, 16px));
-  padding-inline-start: var(--rh-tabs-link-padding-inline-start, var(--rh-space-2xl, 32px));
-  padding-block-end: var(--rh-tabs-link-padding-block-end, var(--rh-space-lg, 16px));
-  padding-inline-end: var(--rh-tabs-link-padding-inline-end, var(--rh-space-2xl, 32px));
-  max-width: 200px;
-  max-height: 100%;
-}
-
 button:focus-visible {
   outline: var(--rh-border-width-md, 2px) solid var(--_button-focus-outline-color);
   outline-offset: -8px;
   border-radius: 10px;
-}
-
-button:before {
-  border: 0 solid transparent;
 }
 
 button:after {
